# Conflicts:
#	untitled/src/Module3/homework/Task3/CollegeStudent.java
#	untitled/src/Module3/homework/Task3/Course.java
#	untitled/src/Module3/homework/Task3/SpecialStudent.java
#	untitled/src/Module3/homework/Task3/Student.java

+++ untitled/src/Module4/homework/Bank.java
@@ -19,7 +19,7 @@ public abstract class Bank {
     //creating abstract methods
     public abstract int getLimitOfWithdrawal();
     public abstract int getLimitOfFunding();
-    public abstract int getMonthlyRate();
+    public abstract double getMonthlyRate();
     public abstract int getCommission();
     public abstract double moneyPaidForSalary();
